Move command to CLM django app and update tests to reflect #11070
